backend_django/requirements.txt
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
requests==2.33.1